Rip out generic route execution
This code was definitely over-engineered and only complicated things.
We keep the route abstraction since this is a useful way to express
potentially complex paths, but exactly how we traverse through the
points of the path is too application dependent. This also allows us to
use Eigen more thoroughly, killing more code.